Thematic 2.0 changes
These functions are now built into Thematic 2.0, they don't break
anything really bad, but it is best to remove them.

You will also need to add back in the 'no-js' class added to the body
that Modernizr hooks into if you use it. This theme doesn't use the
no-js class, but it is probably best to add it.